Moved DrawerItem to adapters.data
