+++ app/src/main/java/com/amaze/filemanager/adapters/DrawerAdapter.java
@@ -42,7 +42,7 @@ import com.amaze.filemanager.filesystem.HybridFile;
 import com.amaze.filemanager.filesystem.Operations;
 import com.amaze.filemanager.filesystem.RootHelper;
 import com.amaze.filemanager.ui.dialogs.GeneralDialogCreation;
-import com.amaze.filemanager.ui.drawer.DrawerItem;
+import com.amaze.filemanager.adapters.data.DrawerItem;
 import com.amaze.filemanager.utils.DataUtils;
 import com.amaze.filemanager.utils.OpenMode;
 import com.amaze.filemanager.utils.Utils;
